+++ CHANGELOG.md
@@ -6,6 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `rich-text` to `search-result` block.
+
+### Fixed
+- Style issues in mobile.
 
 ## [3.26.0] - 2019-08-13