+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [3.28.0] - 2019-08-19
+
 ### Added
 
 - `showProductsCount` prop that controls if the number of products on the page and total quantity of items in a search result is displayed under the `show more` button.